% Generated by roxygen2 (4.0.2): do not edit by hand
\name{calculate_roc}
\alias{calculate_roc}
\title{Calculate the ROC curve}
\usage{
calculate_roc(M, D)
}
\arguments{
\item{M}{continuous marker values or predictions of class labels}

\item{D}{class labels}
}
\value{
Data frame containing cutoffs, and estimated true and false positive fractions
}
\description{
Calculate ROC curve given labels and predictions
}
\examples{
D.ex <- rbinom(100, 1, .5)
calculate_roc(rnorm(100, mean = D.ex), D.ex)
}

