% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_roc.R
\docType{data}
\name{GeomRoc}
\alias{GeomRoc}
\alias{geom_roc}
\title{Empirical Receiver Operating Characteristic Curve}
\format{\preformatted{Classes 'GeomRoc', 'Geom', 'ggproto' <ggproto object: Class GeomRoc, Geom>
    aesthetics: function
    default_aes: uneval
    draw_group: function
    draw_key: function
    draw_layer: function
    draw_panel: function
    extra_params: na.rm
    handle_na: function
    non_missing_aes: size shape
    parameters: function
    required_aes: x y label
    setup_data: function
    use_defaults: function
    super:  <ggproto object: Class Geom> 
}}
\usage{
GeomRoc

geom_roc(mapping = NULL, data = NULL, stat = "roc", n.cuts = 10,
  arrow = NULL, lineend = "butt", linejoin = "round", linemitre = 1,
  alpha.line = 1, alpha.point = 1, size.point = 0.5, labels = TRUE,
  labelsize = 3.88, labelround = 1, na.rm = TRUE, position = "identity",
  show.legend = NA, inherit.aes = TRUE, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{stat}{Use to override the default connection between
\code{geom_roc} and \code{stat_roc}.}

\item{n.cuts}{Number of cutpoints to display along each curve}

\item{arrow}{Arrow specification, as created by \code{\link[grid]{arrow}}}

\item{lineend}{Line end style (round, butt, square)}

\item{linejoin}{Line join style (round, mitre, bevel)}

\item{linemitre}{Line mitre limit (number greater than 1)}

\item{alpha.line}{Alpha level for the lines}

\item{alpha.point}{Alpha level for the cutoff points}

\item{size.point}{Size of cutoff points}

\item{labels}{Logical, display cutoff text labels}

\item{labelsize}{Size of cutoff text labels}

\item{labelround}{Integer, number of significant digits to round cutoff labels}

\item{na.rm}{Remove missing values from curve}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}
}
\description{
Display the empirical ROC curve. Useful for characterizing the classification
accuracy of continuous measurements for predicting binary states
}
\section{Computed variables}{

\describe{
  \item{false_positive_fraction}{estimate of false positive fraction}
  \item{true_positive_fraction}{estimate of true positive fraction}
  \item{cutoffs}{values of m at which estimates are calculated}
}
}

\section{Aesthetics}{

\code{geom_roc} understands the following aesthetics (required aesthetics
are in bold):
\itemize{
  \item \strong{\code{x}} The FPF estimate. This is automatically mapped by \link{stat_roc}
  \item \strong{\code{y}} The TPF estimate. This is automatically mapped by \link{stat_roc}
  smallest level in sort order is assumed to be 0, with a warning
  \item \code{alpha}
  \item \code{color}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
}
}
\examples{
D.ex <- rbinom(250, 1, .5)
rocdata <- data.frame(D = c(D.ex, D.ex),
                   M = c(rnorm(250, mean = D.ex, sd = .4), rnorm(250, mean = D.ex, sd = 1)),
                   Z = c(rep("A", 250), rep("B", 250)))

ggplot(rocdata, aes(m = M, d = D)) + geom_roc()
ggplot(rocdata, aes(m = M, d = D, color = Z)) + geom_roc()
ggplot(rocdata, aes(m = M, d = D)) + geom_roc() + facet_wrap(~ Z)
ggplot(rocdata, aes(m = M, d = D)) + geom_roc(n.cuts = 50)
ggplot(rocdata, aes(m = M, d = D)) + geom_roc(labels = FALSE)
}
\seealso{
See \code{\link{geom_rocci}} for
  displaying rectangular confidence regions for the empirical ROC curve, \code{\link{style_roc}} for
  adding guidelines and labels, and \code{\link{direct_label}} for adding direct labels to the
  curves. Also \link{export_interactive_roc} for creating interactive ROC curve plots for use in a web browser.
}
\keyword{datasets}

