% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_roc.R
\docType{data}
\name{StatRocci}
\alias{StatRocci}
\title{Calculate confidence regions for the empirical ROC curve}
\format{\preformatted{Classes 'StatRocci', 'Stat', 'ggproto' <ggproto object: Class StatRocci, Stat>
    compute_group: function
    compute_layer: function
    compute_panel: function
    default_aes: uneval
    non_missing_aes: 
    parameters: function
    required_aes: m d
    retransform: TRUE
    setup_data: function
    setup_params: function
    super:  <ggproto object: Class Stat> 
}}
\usage{
StatRocci
}
\description{
Confidence intervals for TPF and FPF are calculated using the exact
  method of Clopper and Pearson (1934) each at the level \code{1 - sqrt(1 -
  alpha)}. Based on result 2.4 from Pepe (2003), the cross-product of these
  intervals yields a 1 - alpha % confidence region for (FPF, TPF).
}
\keyword{datasets}

