% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getD3.R
\name{getD3}
\alias{getD3}
\title{Reads included JavaScript functions and returns them as a string for pasting into a webpage}
\usage{
getD3()
}
\description{
Reads included JavaScript functions and returns them as a string for pasting into a webpage
}
\keyword{Internal}

