% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggroc}
\alias{ggroc}
\title{Plot an ROC curve}
\usage{
ggroc(rocdata, fpf_string = "FPF", tpf_string = "TPF")
}
\arguments{
\item{rocdata}{Data frame as returned by calculate_roc}

\item{fpf_string}{Column name identifying false positive fraction}

\item{tpf_string}{Column name identifying true positive fraction}
}
\value{
A ggplot object
}
\description{
Given a result computed by \link{calculate_roc}, plot the curve using ggplot and sensible defaults
}

