% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modCss.R
\name{modCss}
\alias{modCss}
\title{Add custom font size, more customizations to come}
\usage{
modCss(font.size)
}
\arguments{
\item{font.size}{String describing font size, 12px, for example}
}
\description{
Add custom font size, more customizations to come
}
\keyword{Internal}

