% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modJs}
\alias{modJs}
\title{Modify js to use custom name}
\usage{
modJs(selector, prefix)
}
\arguments{
\item{selector}{css selector for the geompoints.object}

\item{prefix}{Prefix of the svg}
}
\description{
Modify js to use custom name
}
\keyword{Internal}

