% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plot_journal_roc}
\alias{plot_journal_roc}
\title{Plot an ROC curve for use in print}
\usage{
plot_journal_roc(ggroc_p, rocdata, font.size = 3, n.cuts = 20)
}
\arguments{
\item{ggroc_p}{An object as returned by \link{ggroc}. It can be modified with
annotations, themes, etc.}

\item{rocdata}{An object as returned by \link{ggroc}. It can be modified with annotations, themes, etc.}

\item{font.size}{Integer that determines font size of cutoff labels}

\item{n.cuts}{Number of cutoffs to display}
}
\value{
A ggplot object
}
\description{
Given a ggroc object, creates a plot suitable for print, with sensible defaults
}

