% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotROC-package.R
\docType{package}
\name{plotROC}
\alias{plotROC}
\alias{plotROC-package}
\title{Tools for plotting ROC Curves}
\description{
Generate Useful ROC Curve Charts for Print and Interactive Use
}
\author{
Michael Sachs (@sachsmc)
}

