% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_prc.R
\name{prc_key}
\alias{prc_key}
\title{Key for PRC geom}
\usage{
prc_key(data, params, size)
}
\arguments{
\item{data}{Data created by stat}

\item{params}{parameters}

\item{size}{Size}
}
\description{
Key for PRC geom
}
\keyword{Internal}
