% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_roc.R
\name{roc_key}
\alias{roc_key}
\title{Key for ROC geom}
\usage{
roc_key(data, params, size)
}
\arguments{
\item{data}{Data created by stat}

\item{params}{parameters}

\item{size}{Size}
}
\description{
Key for ROC geom
}
\keyword{Internal}
