% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shiny_plotROC.R
\name{shiny_plotROC}
\alias{shiny_plotROC}
\title{Start the plotROC Shiny app}
\usage{
shiny_plotROC()
}
\description{
A convenience function to easily start the shiny application. It will open in Rstudio, or in the default web browser.
}

