% Generated by roxygen2 (4.0.2): do not edit by hand
\name{standalone_svg_roc}
\alias{standalone_svg_roc}
\title{Generate a standalone html document displaying an interactive ROC curve}
\usage{
standalone_svg_roc(rocdata, file = NULL, font.size = "12px")
}
\arguments{
\item{rocdata}{An object as returned by \link{ggroc}. It can be modified with annotations, themes, etc.}
}
\value{
NULL opens an interactive document in the browswer
}
\description{
Generate a standalone html document displaying an interactive ROC curve
}

