% Generated by roxygen2 (4.0.2): do not edit by hand
\name{verify_d}
\alias{verify_d}
\title{Check that D is OK}
\usage{
verify_d(D)
}
\arguments{
\item{D}{Vector that will be checked for 2-class labels}
}
\description{
Checks for two classes and gives informative error messages
}
\keyword{Internal}

