% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/style_roc.R
\name{direct_label}
\alias{direct_label}
\title{Add direct labels to a ROC plot}
\usage{
direct_label(ggroc_p, labels = NULL, label.angle = 45, nudge_x = 0,
  nudge_y = 0, size = 6, ...)
}
\arguments{
\item{ggroc_p}{A ggplot object that contains a \link{geom_roc} layer}

\item{labels,}{vector of labels to add directly to the plot next to
the curves. If multiple curves, must be in the same order as the grouping
factor. If NULL, attempts to determine labels from the ggroc_p object}

\item{label.angle}{angle of adjustment for the direct labels}

\item{nudge_x,nudge_y}{Horizontal and vertical adjustment to nudge labels
by. These can be scalars or vectors the same length as the number of labels}

\item{size}{Size of labels}

\item{...}{Other arguments passed to \link[ggplot2]{annotate}}
}
\description{
Add direct labels to a ROC plot
}

