% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_roc.R
\docType{data}
\name{StatRoc}
\alias{StatRoc}
\title{Calculate the empirical ROC curve}
\format{\preformatted{Classes 'StatRoc', 'Stat', 'ggproto' <ggproto object: Class StatRoc, Stat>
    compute_group: function
    compute_layer: function
    compute_panel: function
    default_aes: uneval
    non_missing_aes: 
    parameters: function
    required_aes: m d
    retransform: TRUE
    setup_data: function
    setup_params: function
    super:  <ggproto object: Class Stat> 
}}
\usage{
StatRoc
}
\description{
Calculate the empirical ROC curve
}
\keyword{datasets}

