% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geom_roc.R
\docType{data}
\name{geom_rocci}
\alias{GeomRocci}
\alias{geom_rocci}
\title{Confidence regions for the ROC curve}
\format{\preformatted{Classes 'GeomRocci', 'Geom', 'ggproto' <ggproto object: Class GeomRocci, Geom>
    aesthetics: function
    default_aes: uneval
    draw_group: function
    draw_key: function
    draw_layer: function
    draw_panel: function
    extra_params: na.rm
    handle_na: function
    non_missing_aes: size shape fill
    parameters: function
    required_aes: x y xmin xmax ymin ymax label
    setup_data: function
    use_defaults: function
    super:  <ggproto object: Class Geom> 
}}
\usage{
geom_rocci(mapping = NULL, data = NULL, stat = "rocci",
  position = "identity", show.legend = NA, inherit.aes = TRUE,
  ci.at = NULL, sig.level = 0.05, na.rm = TRUE, ...)

GeomRocci
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link{aes}} or
\code{\link{aes_}}. If specified and \code{inherit.aes = TRUE} (the
default), is combined with the default mapping at the top level of the
plot. You only need to supply \code{mapping} if there isn't a mapping
defined for the plot.}

\item{data}{A data frame. If specified, overrides the default data frame
defined at the top level of the plot.}

\item{position}{Position adjustment, either as a string, or the result of
a call to a position adjustment function.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped.
\code{FALSE} never includes, and \code{TRUE} always includes.}

\item{inherit.aes}{If \code{FALSE}, overrides the default aesthetics,
rather than combining with them. This is most useful for helper functions
that define both data and aesthetics and shouldn't inherit behaviour from
the default plot specification, e.g. \code{\link{borders}}.}

\item{na.rm}{If \code{FALSE} (the default), removes missing values with
a warning.  If \code{TRUE} silently removes missing values.}

\item{...}{other arguments passed on to \code{\link{layer}}. There are
  three types of arguments you can use here:

  \itemize{
  \item Aesthetics: to set an aesthetic to a fixed value, like
     \code{color = "red"} or \code{size = 3}.
  \item Other arguments to the layer, for example you override the
    default \code{stat} associated with the layer.
  \item Other arguments passed on to the stat.
  }}

\item{geom,stat}{Use to override the default connection between
\code{geom_rocci} and \code{stat_rocci}.}

\item{alpha.box}{Alpha level for the confidence regions}

\item{labels}{If TRUE, adds text labels for the cutoffs where the confidence regions are displayed}

\item{labelsize}{Size of cutoff text labels}

\item{labelround}{Integer, number of significant digits to round cutoff labels}
}
\description{
Display rectangular confidence regions for the empirical ROC curve.
}
\section{Aesthetics}{

\code{geom_rocci} understands the following aesthetics (required aesthetics
are in bold). \code{stat_rocci} automatically maps the estimates to the required aesthetics:
\itemize{
  \item \strong{\code{x}} The FPF estimate
  \item \strong{\code{y}} The TPF estimate
  \item \strong{\code{xmin}} Lower confidence limit for the FPF
  \item \strong{\code{xmax}} Upper confidence limit for the FPF
  \item \strong{\code{ymin}} Lower confidence limit for the TPF
  \item \strong{\code{ymax}} Upper confidence limit for the TPF
  \item \code{alpha}
  \item \code{color}
  \item \code{fill}
  \item \code{linetype}
  \item \code{size}
}
}
\examples{
D.ex <- rbinom(250, 1, .5)
rocdata <- data.frame(D = c(D.ex, D.ex),
                   M = c(rnorm(250, mean = D.ex, sd = .4), rnorm(250, mean = D.ex, sd = 1)),
                   Z = c(rep("A", 250), rep("B", 250)))

ggplot(rocdata, aes(m = M, d = D)) + geom_roc() + geom_rocci()
ggplot(rocdata, aes(m = M, d = D, color = Z)) + geom_roc() + geom_rocci()
ggplot(rocdata, aes(m = M, d = D, color = Z)) + geom_roc() + geom_rocci(sig.level = .01)
ggplot(rocdata, aes(m = M, d = D)) + geom_roc(n.cuts = 0) +
geom_rocci(ci.at = quantile(rocdata$M, c(.1, .25, .5, .75, .9)))
ggplot(rocdata, aes(m = M, d = D, color = Z)) + geom_roc() + geom_rocci(linetype = 1)
}
\seealso{
See \code{\link{geom_roc}} for the empirical ROC curve, \code{\link{style_roc}} for
  adding guidelines and labels, and \code{\link{direct_label}} for adding direct labels to the
  curves. Also \link{export_interactive_roc} for creating interactive ROC curve plots for use in a web browser.
}
\keyword{datasets}

