% Generated by roxygen2 (4.0.2): do not edit by hand
\name{ggroc}
\alias{ggroc}
\title{Plot an ROC curve}
\usage{
ggroc(rocdata)
}
\arguments{
\item{rocdata}{Data frame as returned by calculate_roc}
}
\value{
A ggplot object
}
\description{
Given a result computed by \link{calculate_roc}, plot the curve using ggplot and sensible defaults
}

