% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{ggroc}
\alias{ggroc}
\title{Plot an ROC curve}
\usage{
ggroc(
  rocdata,
  fpf_string = "FPF",
  tpf_string = "TPF",
  c_string = "c",
  ci = FALSE,
  label = NULL,
  label.adj.x = 0,
  label.adj.y = 0,
  label.angle = 45,
  plotmath = FALSE,
  xlabel = "False positive fraction",
  ylabel = "True positive fraction"
)
}
\arguments{
\item{rocdata}{Data frame containing true and false positive fractions, and
cutoff values}

\item{fpf_string}{Column name identifying false positive fraction column}

\item{tpf_string}{Column name identifying true positive fraction column}

\item{c_string}{Column name identifying cutoff values}

\item{ci}{Logical, not supported}

\item{label}{Not supported}

\item{label.adj.x}{Not supported}

\item{label.adj.y}{Not supported}

\item{label.angle}{Not supported}

\item{plotmath}{Not supported}

\item{xlabel}{Defaults to "False positive fraction"}

\item{ylabel}{Defaults to "True positive fraction"}
}
\value{
A ggplot object
}
\description{
Deprecated, use \link{geom_roc} instead
}
