% Generated by roxygen2 (4.0.2): do not edit by hand
\name{modJs}
\alias{modJs}
\title{Modify javascript function to use custom name}
\usage{
modJs(selector, prefix, rect = "")
}
\arguments{
\item{selector}{css selector for the geompoints.object}

\item{prefix}{Prefix of the svg}

\item{rect}{If not empty, the selector for the geom_rects for the confidence
  regions}
}
\description{
Modify javascript function to use custom name
}
\keyword{Internal}

