% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/ggroc.R
\name{multi_ggroc}
\alias{multi_ggroc}
\title{Plot multiple ROC curves}
\usage{
multi_ggroc(datalist, fpf_string = rep("FPF", length(datalist)),
  tpf_string = rep("TPF", length(datalist)), label = NULL,
  label.adj.x = rep(0, length(datalist)), label.adj.y = rep(0,
  length(datalist)), label.angle = rep(45, length(datalist)),
  lty = 1:length(datalist), color = NULL, plotmath = TRUE,
  legend = FALSE, ...)
}
\arguments{
\item{datalist}{List of data frames each containing true and false positive
fractions and cutoffs}

\item{fpf_string}{Column names identifying false positive fraction}

\item{tpf_string}{Column names identifying true positive fraction}

\item{label}{Optional vector of direct labels for the ROC curve, same length
as \code{datalist}}

\item{label.adj.x}{Adjustment for the positioning of the label, same length
as \code{datalist}}

\item{label.adj.y}{Adjustment for the positioning of the label, same length
as \code{datalist}}

\item{label.angle}{Adjustment for angle of label, same length as
\code{datalist}}

\item{lty}{Vector of positive integers denoting line types to distinguish curves}

\item{color}{Optional vector of colors to distinguish curves. If NULL, all curves will be black.}

\item{plotmath}{Logical. If TRUE, labels will be parsed as expressions. See \code{?plotmath} for details.}

\item{legend}{Logical. If TRUE, legend with labels is plotted in lower right corner of plot.}

\item{...}{Other aethetics settings passed to \code{geom_path}, e.g. \code{lwd} or \code{size}.}
}
\value{
A ggplot object
}
\description{
Given a list of results computed by \link{calculate_roc}, plot the curve
using ggplot with sensible defaults. Pass the resulting object and data to
\link{export_interactive_roc}, \link{plot_interactive_roc}, or
\link{plot_journal_roc}.
}

