% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multi_ggroc}
\alias{multi_ggroc}
\title{Plot multiple ROC curves}
\usage{
multi_ggroc(datalist, fpf_string = rep("FPF", length(datalist)),
  tpf_string = rep("TPF", length(datalist)), lty = 1:length(datalist))
}
\arguments{
\item{datalist}{List of data frames as returned by calculate_roc}

\item{fpf_string}{Column names identifying false positive fraction}

\item{tpf_string}{Column names identifying true positive fraction}

\item{lty}{Line types to distinguish curves}
}
\value{
A ggplot object
}
\description{
Given a list of results computed by \link{calculate_roc}, plot the curve using ggplot and sensible defaults
}

