% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_roc.R
\name{plot_interactive_roc}
\alias{plot_interactive_roc}
\title{Generate a standalone html document displaying an interactive ROC curve}
\usage{
plot_interactive_roc(ggroc, file = NULL, ...)
}
\arguments{
\item{ggroc}{An object as returned by \link{ggroc} or \link{multi_ggroc}. It can be modified with annotations, themes, etc.}

\item{file}{A path to save the result to. If NULL, will save to a temporary directory}

\item{...}{arguments passed to \link{export_interactive_roc}}
}
\value{
NULL opens an interactive document in Rstudio or the default web browser
}
\description{
Generate a standalone html document displaying an interactive ROC curve
}

