% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/deprecated.R
\name{plot_journal_roc}
\alias{plot_journal_roc}
\title{Plot an ROC curve for use in print}
\usage{
plot_journal_roc(ggroc_p, font.size = 3, n.cuts = 20, ci.at = NULL,
  opacity = 0.3, lty = NULL, color = NULL, lwd = NULL, legend = FALSE)
}
\arguments{
\item{ggroc_p}{An object as returned by \link{ggroc} or \link{multi_ggroc}.
It can be modified with annotations, themes, etc.}

\item{font.size}{Not supported}

\item{n.cuts}{Not supported}

\item{ci.at}{Not supported}

\item{opacity}{Not supported}

\item{lty}{Not supported}

\item{color}{Not supported}

\item{lwd}{Not supported}

\item{legend}{Not supported}
}
\value{
A ggplot object
}
\description{
Deprecated, use \link{style_roc} instead
}

