% Generated by roxygen2 (4.0.2): do not edit by hand
\name{svgize_roc}
\alias{svgize_roc}
\title{Generate svg code for an ROC curve object. Includes necessary styling and
Javascript code}
\usage{
svgize_roc(ggroc_p, cutoffs, font.size = "12px", prefix = "a")
}
\arguments{
\item{ggroc_p}{An object as returned by \link{ggroc}. It can be modified with
annotations, themes, etc.}

\item{cutoffs}{Vector of cutoff values}

\item{font.size}{Character string that determines font size of cutoff labels}

\item{prefix}{A string to assign to the objects within the svg. Enables
  unique idenfication by the javascript code}
}
\value{
A string containing the svg code necessary to plot the ROC curve in a
  browser
}
\description{
If you intend to include more than one of these objects in a single page, use
a different \code{prefix} string for each one.
}

