% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_roc.R
\name{verify_d}
\alias{verify_d}
\title{Check that D is suitable for using as binary disease status}
\usage{
verify_d(D)
}
\arguments{
\item{D}{Vector that will be checked for 2-class labels}
}
\description{
Checks for two classes and gives a warning message indicating which level is assumed to be 0/1
}
\keyword{Internal}

