% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/calculate_roc.R
\name{verify_d}
\alias{verify_d}
\title{Check that D is OK for using as binary disease status}
\usage{
verify_d(D)
}
\arguments{
\item{D}{Vector that will be checked for 2-class labels}
}
\description{
Checks for two classes and gives informative error messages
}
\keyword{Internal}

